Moved redux folder out of app directory
