+++ __tests__/test-utils.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import { render, RenderOptions } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import store from '@/app/Redux/store'
+import React from "react";
+import { render, RenderOptions } from "@testing-library/react";
+import { Provider } from "react-redux";
+import store from "@/Redux/store";
 
 interface CustomRenderOptions extends RenderOptions {
   initialState?: any;
@@ -14,5 +14,5 @@ const customRender = (
   return render(<Provider store={store}>{ui}</Provider>, renderOptions);
 };
 
-export * from '@testing-library/react';
+export * from "@testing-library/react";
 export { customRender as render };
