+++ src/app/StoreProvider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Provider } from "react-redux";
-import store from "./Redux/store";
+import store from "../Redux/store";
 import React from "react";
 
 const StoreProvider = ({ children }: { children: React.ReactNode }) => {
